Fix Pattern tests
